+++ activerecord/lib/active_record/associations/has_one_association.rb
@@ -38,7 +38,7 @@ module ActiveRecord
             when :destroy
               target.destroy
             when :nullify
-              target.update_column(reflection.foreign_key, nil)
+              target.update_attribute(reflection.foreign_key, nil)
           end
         end
       end